move custom errors to utils
